rgw: move RGWObjManifest stuff into rgw_manifest.cc,h
