+++ src/rgw/CMakeLists.txt
@@ -74,6 +74,7 @@ set(librgw_common_srcs
   rgw_ldap.cc
   rgw_lc.cc
   rgw_lc_s3.cc
+  rgw_manifest.cc
   rgw_metadata.cc
   rgw_multi.cc
   rgw_multi_del.cc
